fix: add interactive chart tooltips, fix deactivate button UI, and resolve cross-feature import violations
