+++ src/features/management-report/components/StageDelayView.tsx
@@ -16,7 +16,7 @@ import {
 import { Button } from '@/shared/components/ui/button'
 import { AlertCircle, RefreshCw, Layers, AlertTriangle } from 'lucide-react'
 import { cn } from '@/shared/lib/utils'
-import { formatDuration } from '@/features/bed-dashboard/lib/duration-formatters'
+import { formatDuration } from '@/shared/lib/duration-formatters'
 import { useStageDelayData, PRESETS } from '../hooks/useStageDelayData'
 import { StageDelayBarChart } from './StageDelayBarChart'
 import { StageDelayTable } from './StageDelayTable'
